+++ src/imio/pm/wsclient/testing.py
@@ -18,7 +18,7 @@ WS4PMCLIENT_ZCML = zca.ZCMLSandbox(filename="testing.zcml",
 WS4PMCLIENT_Z2 = z2.IntegrationTesting(bases=(z2.STARTUP, WS4PMCLIENT_ZCML),
                                        name='WS4PMCLIENT_Z2')
 
-WS4PMCLIENT = PloneWithPackageLayer(
+WS4PMCLIENT = WSCLIENTLayer(
     zcml_filename="testing-settings.zcml",
     zcml_package=imio.pm.wsclient,
     additional_z2_products=('imio.dashboard',